feat: Adding Jobs History API - Adding Jobs Controller and Service, which calls AuditsService to get the history for the jobs feature - Swapping endpoints in UI - Fixing issue in CellDateTime when there's no value for the cell - Doing some more organization cleanup with respect to features - Removing unused component
