+++ packages/ui/src/api.ts
@@ -1,4 +1,4 @@
-﻿import { ApplicationsApi, AuditsApi, CompaniesApi, Configuration } from "@incutonez/life-stats-spec";
+﻿import { ApplicationsApi, AuditsApi, CompaniesApi, Configuration, JobsApi } from "@incutonez/life-stats-spec";
 
 export const apiConfig = new Configuration({
 	basePath: import.meta.env.VITE_BASE_API || `http://${location.hostname}:3000`,
@@ -6,6 +6,8 @@ export const apiConfig = new Configuration({
 		headers: {},
 	},
 });
+
+export const JobsAPI = new JobsApi(apiConfig);
 
 export const ApplicationsAPI = new ApplicationsApi(apiConfig);
