+++ packages/ui/src/composables/applications.ts
@@ -3,7 +3,7 @@ import { type ApplicationViewModel, EnumApplicationStatus } from "@incutonez/lif
 import { useMutation, useQuery, useQueryClient } from "@tanstack/vue-query";
 import clone from "just-clone";
 import { ApplicationsAPI } from "@/api.ts";
-import { QueryGetApplication, QueryListApplications, QueryListAudits, QueryListCompanies } from "@/constants.ts";
+import { QueryGetApplication, QueryListApplications, QueryListCompanies, QueryListJobAudits } from "@/constants.ts";
 import { RouteCreate } from "@/router/index.ts";
 import { setApplicationRecords } from "@/stores/applications.ts";
 import { useAppDispatch } from "@/stores/main.ts";
@@ -83,7 +83,7 @@ export function useDeleteApplication() {
 				queryKey: [QueryListCompanies],
 			});
 			await queryClient.invalidateQueries({
-				queryKey: [QueryListAudits],
+				queryKey: [QueryListJobAudits],
 			});
 		},
 	});
